backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
databases
python-jose[cryptography]
passlib[bcrypt]
PyJWT
python-decouple
python-multipart
python-dotenv